Rename file to match class (Web3 -> Web3Loader)
