+++ web-app/src/components/web3/__mocks__/Web3Loader.js
